+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Unreleased]
 ### Fixed
 - Cleaned up ranged attack chat cards.
+- No longer possible for quality armor to provide a positive defense bonus.
 
 ## [0.3.5] - 2022-12-18
 ### Added